app/requirements.txt
faiss-cpu
openai
fastapi
uvicorn
langchain
langchain-community
langchain-openai
tiktoken
sqlalchemy
